code reorganization pre call options impl
